+++ src/app/shared/NoteContentKind1985.svelte
@@ -1,13 +1,13 @@
 <script lang="ts">
-  import {isNil} from "ramda"
-  import {getLabelQuality} from "src/util/nostr"
+  import {isNil} from 'ramda'
+  import {getRating} from "src/util/nostr"
   import NoteContentLabel from "src/app/shared/NoteContentLabel.svelte"
   import NoteContentRating from "src/app/shared/NoteContentRating.svelte"
   import NoteContentKind1 from "src/app/shared/NoteContentKind1.svelte"
 
   export let note, anchorId, maxLength, showEntire
 
-  const rating = getLabelQuality("review/relay", note)
+  const rating = getRating(note)
 </script>
 
 <div class="flex flex-col gap-2 overflow-hidden text-ellipsis">