+++ src/util/nostr.ts
@@ -1,9 +1,10 @@
 import {Tags} from "paravel"
+import {last} from 'ramda'
 import {nip19} from "nostr-tools"
-import {pick, is, mergeLeft, last, identity} from "ramda"
+import {pick, is, mergeLeft, identity} from "ramda"
 import {between, avg} from "hurdak"
 import type {Filter, Event} from "src/engine"
-import {tryJson} from "src/util/misc"
+import {tryJson} from 'src/util/misc'
 
 export const noteKinds = [1, 30023, 1063, 9802, 1808, 32123]
 export const personKinds = [0, 2, 3, 10000, 10002]
@@ -41,14 +42,22 @@ export const toHex = (data: string): string | null => {
 export const mergeFilter = (filter: Filter | Filter[], extra: Filter) =>
   is(Array, filter) ? filter.map(mergeLeft(extra)) : {...filter, ...extra}
 
-export const getLabelQuality = (label: string, event: Event) => {
-  const json = tryJson(() => JSON.parse(last(Tags.from(event).type("l").equals(label).first())))
+export const getRating = (event: Event) => {
+  const tags = Tags.from(event)
+  const rating = parseInt(tags.nthEq(2, "rating").values().first())
+
+  if (!isNaN(rating)) {
+    return rating
+  }
+
+  const json = tryJson(() => {
+    return JSON.parse(last(tags.nthEq(1, "review/relay").first()))
+  })
 
   return (json as {quality?: number})?.quality
 }
 
-export const getAvgQuality = (label: string, events: Event[]) =>
-  avg(events.map(e => getLabelQuality(label, e)).filter(identity))
+export const getAvgRating = (events: Event[]) => avg(events.map(getRating).filter(identity))
 
 export const isHex = x => x.match(/^[a-f0-9]{64}$/)